requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
albumentations==1.3.1
alembic==1.12.1
async-timeout==4.0.3
asynctest==0.13.0
attrs==24.2.0
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
click==8.1.8
cloudpickle==2.2.1
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
databricks-cli==0.18.0
docker==6.1.3
entrypoints==0.4
et-xmlfile==1.1.0
Flask==2.2.5
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
future @ file:///croot/future_1677599870788/work
gitdb==4.0.12
GitPython==3.1.44
grad-cam==1.4.6
greenlet==3.1.1
gunicorn==20.1.0
idna @ file:///croot/idna_1666125576474/work
imageio==2.31.2
imgaug==0.4.0
importlib-metadata==5.2.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.6
joblib==1.3.2
kiwisolver==1.4.5
lightning-utilities==0.10.1
Mako==1.2.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mkl-fft==1.3.1
mkl-random @ file:///tmp/build/80754af9/mkl_random_1626179032232/work
mkl-service==2.4.0
mlflow==1.30.1
multidict==6.0.5
networkx==2.6.3
numpy @ file:///opt/conda/conda-bld/numpy_and_numpy_base_1653915516269/work
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
opencv-python==4.4.0.46
opencv-python-headless==4.2.0.34
openpyxl==3.1.3
packaging==21.3
pandas==1.3.5
Pillow==9.4.0
prometheus-client==0.17.1
prometheus_flask_exporter==0.23.2
protobuf==4.24.4
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
PyJWT==2.8.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.1.4
PySocks @ file:///tmp/build/80754af9/pysocks_1594394576006/work
python-dateutil==2.9.0.post0
pytorch-lightning==1.8.6
pytz==2022.7.1
PyWavelets==1.3.0
PyYAML @ file:///croot/pyyaml_1670514731622/work
qudida==0.0.4
querystring-parser==1.2.4
requests @ file:///opt/conda/conda-bld/requests_1657734628632/work
scikit-image==0.19.3
scikit-learn==1.0.2
scipy==1.7.3
seaborn==0.12.2
shapely==2.0.7
six @ file:///tmp/build/80754af9/six_1644875935023/work
smmap==5.0.2
SQLAlchemy==1.4.54
sqlparse==0.4.4
tabulate==0.9.0
tensorboardX==2.6.2.2
termcolor==1.1.0
threadpoolctl==3.1.0
tifffile==2021.11.2
timm==0.4.12
torch==1.13.1
torchaudio==0.13.1
torchcam==0.3.2
torchmetrics==0.11.4
torchsampler==0.1.2
torchvision==0.14.1
tqdm==4.67.1
ttach==0.0.3
typing_extensions @ file:///croot/typing_extensions_1669924550328/work
urllib3 @ file:///croot/urllib3_1673575502006/work
websocket-client==1.6.1
Werkzeug==2.2.3
yacs==0.1.8
yarl==1.9.4
zipp==3.15.0
